// lib/api/model/studentAttendanceOverview.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ModuleAttendanceSummary } from './moduleAttendanceSummary';

export interface StudentAttendanceOverview {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  totalModules: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  presentCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  lateCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  absentCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  excusedCount: number | string;
  modules: ModuleAttendanceSummary[];
}